requirements.unittest.txt
alembic==1.0.1
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.2
Babel==2.6.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
cryptography==2.3.1
filelock==3.0.10
Flask==1.0.2
Flask-Babel==0.12.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
PyMySQL==0.9.2
pytest==4.0.2
python-dateutil==2.7.4
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.6
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.12
toml==0.10.0
tox==3.9.0
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==16.5.0
Werkzeug==0.14.1
wrapt==1.10.11
